+++ tslib/src/functions/check-allocator/ICheckAllocatorDataService.ts
@@ -8,6 +8,7 @@ export interface ICheckAllocationDataService {
    */
   getPupilsBySchoolUuid (schoolUUID: string): Promise<Array<IPupil>>
 }
+
 export class CheckAllocationDataService implements ICheckAllocationDataService {
   async getPupilsBySchoolUuid (schoolUUID: string): Promise<Array<IPupil>> {
     throw new Error('Method not implemented.')
