+++ tslib/src/functions/check-allocator/ICheckFormAllocationService.ts
@@ -1,6 +1,7 @@
 export interface ICheckFormAllocationService {
   allocate (pupilId: number): Promise<any>
 }
+
 export class CheckFormAllocationService implements ICheckFormAllocationService {
   allocate (pupilId: number): Promise<any> {
     return Promise.reject(new Error('not implemented'))
