+++ tslib/tsconfig.json
@@ -77,6 +77,6 @@
   ],
   "exclude": [
     "**/*.spec.ts",
-    "src/sql/sql.service.ts"
+    "src/sql/sql.service.converted.ts"
 ],
 }
